+++ user-interface/src/styles/breackpoints.js
@@ -1,0 +1,30 @@
+// Breakpoints:
+export const bkpColumnToRow = {
+  base: "column",
+  sm: "column",
+  md: "column",
+  lg: "row",
+  xl: "row",
+};
+
+export const bkpCentralPanel = {
+  base: "98%",
+  sm: "98%",
+  md: "55%",
+  lg: "45%",
+  xl: "40%",
+};
+
+export const bkpBigImage = {
+  base: 170,
+  sm: 170,
+  md: 200,
+  lg: 250,
+  xl: 400,
+};
+
+export const bkpHSeparator = {
+  base: 10,
+  sm: 10,
+  md: 10,
+};
